requirements.txt
sentence-transformers==4.1.0
streamlit==1.44.1
pandas==2.2.3
jira==3.8.0
python-dotenv==1.1.0
ipywidgets==8.1.6
